:recycle: refactor: Passando cookies do usuário para o menu através de parâmetro
