+++ pages/company/index.tsx
@@ -1,0 +1,6 @@
+export default function Company() {
+    return (
+        <>
+        </>
+    )
+}
